# break up selecting keys from decorating in new keys