--- a/README.md
+++ b/README.md
@@ -3,7 +3,9 @@
 A JavaScript module for decoding [Mapbox vector tiles](https://github.com/mapbox/vector-tile-spec),
 doing various modifications in JS and reencoding back. It can:
 
-- add and remove specific properties in a layer (`decorateLayer`)
+- add new properties to a layer (`decorateLayer`)
+- filter out features that don't have specific properties (`filterByKeys`)
+- select properties to keep on each feature (`selectKeys`)
 - merge features with the same properties and type into one, and sort geometries within one feature by proximity for better compression (`mergeLayer`)
 - get all values of a specific property in a layer (`getLayerValues`)
 - rename layers (`layer.name = 'foo'`)