giving functions smaller surface area of responsibility - this gives us more control 💪 

Now, instead of decorateLayer doing key selection + decorating, `selectKeys` is it's own method, with `decorateLayer` handling only the process of applying new keys

old: 

```js
// decorateLayer(layer, keysToKeep, newProps)
Decorator.decorateLayer(layer, ['id', 'type'], [/* ... new properties... */]);
```

new: 
```js
// selectKeys(layer, keysToKeep)
Decorator.selectKeys(layer, ['id', 'type']);
// decorateLayer(layer, newProps)
Decorator.decorateLayer(layer, [/* ... new properties ... */]);
```

This requires implementing clients to do a little extra work but affords more control over what happens and when it happens. For example, in the old setup, it's impossible to decorate a layer with new properties and then specify "keys to keep" on the output side (key selection always applied to the source data. `newProps` was never filtered). It unlocks more fine-grained controls, e.g https://github.com/mapbox/tilelive-decorator/issues/10

This does require an extra loop over `layer.features` to achieve the same work, so I've tested performance indirectly via tilelive-decorator benchmarks;

old: 

```js
decorator x 235 ops/sec ±4.45% (34 runs sampled)
decorator#requiredKeys x 599 ops/sec ±4.86% (35 runs sampled)
```

new: 

```js
decorator x 220 ops/sec ±6.46% (35 runs sampled)
decorator#requiredKeys x 565 ops/sec ±5.31% (35 runs sampled)
```

IMO this is fine, and an even bigger drop would be acceptable to me for the extra 💪  this gets us. I'd like to do a test of this in some real world infrastructure to see how it runs at scale before 🚢 , but I'm happy so far.