@tcql 2017-04-04
real world performance looks good 👍