@mourner 2017-04-04
> it's impossible to decorate a layer with new properties and then specify "keys to keep" on the output side (key selection always applied to the source data. newProps was never filtered)

Can you give a few more examples of how this would be used in the real world? E.g. with the example above, would you `decorateLayer` first and then `selectKeys` instead of the other way around?