@tcql 2017-04-04
> would you decorateLayer first and then selectKeys instead of the other way around?

@mourner yes exactly. The example in mind is mapbox traffic, which applies records from redis (which may include some extra properties we don't want on the output tile) to a road graph that includes many extra properties. After decorating, we want to specify strictly which properties should exist on the output.