@mourner 2017-04-04
@tcql would you achieve the same result by just generating a newProps array with filtered keys first and only then decorating?