@tcql 2017-04-04
@mourner hmm yes, that'd be possible. This functionality, however, is more flexible since you don't really need to know where properties came from (new props? or from the source tile?) in order to filter them. 

I liked this solution because specifying keepKeys as part of `decorateLayer` always felt weird to me - adding new properties and filtering what keys to keep are two separate pieces of work, and it's entirely possible you could want to add new keys without specifying which keys to keep at all (just everything in the source + everything that came from redis).