@tcql 2017-04-04
While we're renaming things, we should work out better names for `filterByKeys` and `selectKeys` - currently, the established pattern is for methods to be suffixed with `Layer` when they operate on a single layer as opposed to a tile object: 

- `getLayerValues`
- `decorateLayer`
- `mergeLayer`

so we should probably either follow that pattern or ditch it entirely